Fixed Small Bug in Readme.md
I fixed a small bug in the Readme.md about the release of the code. The new version explains this in more detail.
In addition, I have added a clarification of the licenses, which ensures that the code of derived works must also be published. Since the European Union Public License is available in all official languages of the European Union, this is the preferred option.